python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
psycopg2-binary==2.9.9
supabase==2.0.2
openai==1.50.0